refactor: move coverage envelope contracts to output crate
