+++ crates/cli/src/output_envelope.rs
@@ -13,7 +13,10 @@ use std::sync::atomic::{AtomicBool, Ordering};
 pub use fallow_output::{
     CheckGroupedEntry, CheckGroupedOutput, CheckOutput, CheckOutputInput, CodeClimateIssue,
     CodeClimateIssueKind, CodeClimateLines, CodeClimateLocation, CodeClimateOutput,
-    CodeClimateSeverity, DupesOutput, DupesOutputInput, GroupByMode, HealthOutput,
+    CodeClimateSeverity, CoverageAnalyzeOutput, CoverageAnalyzeSchemaVersion,
+    CoverageSetupFileToEdit, CoverageSetupFramework, CoverageSetupMember, CoverageSetupOutput,
+    CoverageSetupPackageManager, CoverageSetupRuntimeTarget, CoverageSetupSchemaVersion,
+    CoverageSetupSnippet, DupesOutput, DupesOutputInput, GroupByMode, HealthOutput,
     HealthOutputInput, WorkspaceDiagnosticOutput, apply_config_fixable_to_duplicate_exports,
     build_check_output, build_check_summary, build_dupes_output, build_health_output,
 };
@@ -22,7 +25,7 @@ use fallow_types::output::NextStep;
 use serde::Serialize;
 
 use crate::audit::{AuditAttribution, AuditSummary, AuditVerdict};
-use crate::health_types::{HealthGroup, HealthReport, RuntimeCoverageReport};
+use crate::health_types::{HealthGroup, HealthReport};
 use crate::output_dupes::DupesReportPayload;
 use crate::report::dupes_grouping::DuplicationGroup;
 
@@ -129,99 +132,6 @@ pub fn apply_root_kind(value: &mut serde_json::Value, kind: &'static str) {
         );
     }
 }
-/// `fallow coverage setup --json` envelope.
-#[derive(Debug, Clone, Serialize)]
-#[cfg_attr(feature = "schema", derive(schemars::JsonSchema))]
-#[cfg_attr(feature = "schema", schemars(title = "fallow coverage setup --json"))]
-pub struct CoverageSetupOutput {
-    pub schema_version: CoverageSetupSchemaVersion,
-    pub framework_detected: CoverageSetupFramework,
-    pub package_manager: Option<CoverageSetupPackageManager>,
-    pub runtime_targets: Vec<CoverageSetupRuntimeTarget>,
-    pub members: Vec<CoverageSetupMember>,
-    pub config_written: Option<serde_json::Value>,
-    pub commands: Vec<String>,
-    pub files_to_edit: Vec<CoverageSetupFileToEdit>,
-    pub snippets: Vec<CoverageSetupSnippet>,
-    pub dockerfile_snippet: Option<String>,
-    pub next_steps: Vec<String>,
-    pub warnings: Vec<String>,
-    #[serde(rename = "_meta", default, skip_serializing_if = "Option::is_none")]
-    pub meta: Option<serde_json::Value>,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize)]
-#[cfg_attr(feature = "schema", derive(schemars::JsonSchema))]
-pub enum CoverageSetupSchemaVersion {
-    #[serde(rename = "1")]
-    V1,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize)]
-#[cfg_attr(feature = "schema", derive(schemars::JsonSchema))]
-#[serde(rename_all = "snake_case")]
-pub enum CoverageSetupFramework {
-    #[serde(rename = "nextjs")]
-    NextJs,
-    #[serde(rename = "nestjs")]
-    NestJs,
-    Nuxt,
-    #[serde(rename = "sveltekit")]
-    SvelteKit,
-    Astro,
-    Remix,
-    Vite,
-    PlainNode,
-    Unknown,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize)]
-#[cfg_attr(feature = "schema", derive(schemars::JsonSchema))]
-#[serde(rename_all = "lowercase")]
-pub enum CoverageSetupPackageManager {
-    Npm,
-    Pnpm,
-    Yarn,
-    Bun,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize)]
-#[cfg_attr(feature = "schema", derive(schemars::JsonSchema))]
-#[serde(rename_all = "lowercase")]
-pub enum CoverageSetupRuntimeTarget {
-    Node,
-    Browser,
-}
-
-#[derive(Debug, Clone, Serialize)]
-#[cfg_attr(feature = "schema", derive(schemars::JsonSchema))]
-pub struct CoverageSetupMember {
-    pub name: String,
-    pub path: String,
-    pub framework_detected: CoverageSetupFramework,
-    pub package_manager: Option<CoverageSetupPackageManager>,
-    pub runtime_targets: Vec<CoverageSetupRuntimeTarget>,
-    pub files_to_edit: Vec<CoverageSetupFileToEdit>,
-    pub snippets: Vec<CoverageSetupSnippet>,
-    pub dockerfile_snippet: Option<String>,
-    pub warnings: Vec<String>,
-}
-
-#[derive(Debug, Clone, Serialize)]
-#[cfg_attr(feature = "schema", derive(schemars::JsonSchema))]
-pub struct CoverageSetupFileToEdit {
-    pub path: String,
-    pub reason: String,
-}
-
-#[derive(Debug, Clone, Serialize)]
-#[cfg_attr(feature = "schema", derive(schemars::JsonSchema))]
-pub struct CoverageSetupSnippet {
-    pub label: String,
-    pub path: String,
-    pub content: String,
-}
-
 /// `fallow audit --format json` envelope.
 #[derive(Debug, Clone, Serialize)]
 #[cfg_attr(feature = "schema", derive(schemars::JsonSchema))]
@@ -309,28 +219,6 @@ pub struct CombinedMeta {
     pub health: Option<Meta>,
     #[serde(default, skip_serializing_if = "Option::is_none")]
     pub telemetry: Option<TelemetryMeta>,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize)]
-#[cfg_attr(feature = "schema", derive(schemars::JsonSchema))]
-pub enum CoverageAnalyzeSchemaVersion {
-    #[serde(rename = "1")]
-    V1,
-}
-
-#[derive(Debug, Clone, Serialize)]
-#[cfg_attr(feature = "schema", derive(schemars::JsonSchema))]
-#[cfg_attr(
-    feature = "schema",
-    schemars(title = "fallow coverage analyze --format json")
-)]
-pub struct CoverageAnalyzeOutput {
-    pub schema_version: CoverageAnalyzeSchemaVersion,
-    pub version: ToolVersion,
-    pub elapsed_ms: ElapsedMs,
-    pub runtime_coverage: RuntimeCoverageReport,
-    #[serde(rename = "_meta", default, skip_serializing_if = "Option::is_none")]
-    pub meta: Option<Meta>,
 }
 
 /// Envelope emitted by `fallow explain <issue-type> --format json`.
